Add files via upload
Overview of our project: It shows all the product in the store.It takes user input as whether they want to purchase products or they want to add product to their store.   By referring to product details the user can select the product.The user have to enter their details in the beginning, and then have to enter how many product id they want to purchase and  quantity of product . As they enter these details their bill amount will be calculated. Then user has to enter yes if they have completed the payment.Then at the end it displays the transaction details of the user which includes the products purchased and its price.